DRAFT: RESTful API helpers
