+++ apps/astro-app/src/utils/data.ts
@@ -76,51 +76,6 @@ export function handleRecordOn(
   });
 }
 
-/**
- * Promise wrapper for dataset fetching.
- * @param fetchPromise
- * @param schema The zod schema to validate against.
- * @returns A promise to fetch the data.
- */
-export function safeFetchDataset<T extends ZodType<any>>(
-  fetchPromise: Promise<Response>,
-  schema: T,
-): Promise<z.infer<T>> {
-  return new Promise((resolve, reject) => {
-    fetchPromise
-      .then((response) => {
-        if (!response.ok) {
-          reject(
-            `Server response not OK: ${response.status} ${response.statusText}`,
-          );
-          return;
-        }
-
-        response
-          .json()
-          .then((body) => {
-            const result = schema.safeParse(body);
-            if (!result.success) {
-              reject(result.error);
-              return;
-            }
-
-            if (!result.data) {
-              // if Zod's behaviour is unexpected,
-              reject(
-                "Undefined data? Contact website administrator or dev for a fix.",
-              );
-              return;
-            } else {
-              resolve(result.data);
-            }
-          })
-          .catch(reject);
-      })
-      .catch(reject);
-  });
-}
-
 export function getFallbackValue(datatype: "int" | "integer"): number;
 export function getFallbackValue(datatype: "float" | "number"): number;
 export function getFallbackValue(
@@ -288,109 +243,6 @@ export interface formOutput {
   data: Record<string, any>;
   tags?: String[];
   descriptors?: Record<string, any>;
-}
-
-/**
- * Serializes data bound for a database.
- * @param values The values that need to be serialized.
- * @param schema The schema that relates to the given values.
- * @returns An object filled with the formatted values, or a string with a corresponding error message if the schema does not match the values.
- */
-function serializeDataObject(
-  values: Record<string, any>,
-  schema: Array<DataColumn>,
-): Record<string, any> {
-  let formattedValues: Record<string, any> = {};
-
-  for (let column_schema of schema) {
-    // check the comments first
-    if (column_schema.comments) {
-      formattedValues[`${column_schema.name}_comments`] =
-        values[`${column_schema.name}_comments`] ?? "''";
-    }
-
-    // check if the user really wanted to submit that or not
-    if (values[column_schema.name] == null) {
-      continue;
-    }
-
-    // single-quote strings, dates, times, and timestamps
-    switch (column_schema.datatype) {
-      case "string":
-      case "str":
-      case "text":
-        formattedValues[column_schema.name] = `'${values[column_schema.name]}'`;
-        break;
-      case "date":
-      case "time":
-      case "timestamp":
-        formattedValues[column_schema.name] =
-          `'${typeof values[column_schema.name] == "string" ? values[column_schema.name] : values[column_schema.name].toISOString()}'`.replace(
-            "Z",
-            "",
-          );
-        break;
-      default:
-        formattedValues[column_schema.name] = values[column_schema.name];
-        break;
-    }
-  }
-
-  return formattedValues;
-}
-
-/**
- * Attempts to serialize database-bound data.
- * @param data The data to serialize.
- * @param strict Whether or not to throw out when a schema is incomplete (partially cooperative)
- * @param tableInfo The table schema of the data.
- */
-export function serializeFormOutput(
-  data: formOutput,
-  strict: boolean,
-  tableInfo: TableInfo,
-): formOutput {
-  let serializedValues: formOutput = {
-    data: serializeDataObject(data.data, tableInfo.schema),
-  };
-
-  if ("tags" in data && data["tags"]) {
-    if ("primary_tag" in data.data) {
-      throw "Primary tag already set? Contact dev for a fix.";
-    }
-
-    data.data["primary_tag"] = data.tags.shift();
-
-    if (strict && !data.data["primary_tag"]) {
-      throw "Primary tag cannot be empty.";
-    }
-
-    serializedValues["tags"] = data.tags;
-  }
-
-  if ("descriptors" in data && data.descriptors != undefined) {
-    serializedValues["descriptors"] = {};
-
-    // format every descriptor type.
-    for (const descriptor_schema of tableInfo.descriptors)
-      serializedValues.descriptors[descriptor_schema.name] = data.descriptors[
-        descriptor_schema.name
-      ].map((value: { [x: string]: any }) =>
-        serializeDataObject(value, descriptor_schema.schema),
-      );
-  }
-
-  return serializedValues;
-}
-
-// export function deserializedDatabaseRecord(data: Record<string, unknown>) {}
-
-/**
- * Checks if tagging is enabled on the given table.
- * @param tableInfo The schema of the table to check.
- */
-export function isTaggingEnabled(tableInfo: TableInfo): boolean {
-  return tableInfo.tagging;
 }
 
 export { parseDatabaseValue };
